[Xamarin.Android.Build.Tasks] rename AndroidUpdateResDir to AndroidComputeResPaths.
The `AndroidUpdateResDir.cs` source file contained two
classes. `AndroidComputeResPaths` and `RemoveUnknownFiles`.
Which never really made any sense.

This PR moves the two classes in the their own correctly
named files.
